Renaming process -> score, cleaning up a bit
